+++ PicCharge/PicCharge/Global/Extensions/URL+.swift
@@ -1,0 +1,18 @@
+//
+//  URL+.swift
+//  PicCharge
+//
+//  Created by 남유성 on 1/21/25.
+//
+
+import Foundation
+
+extension URL {
+    var queryItems: [String: String]? {
+        guard let components = URLComponents(url: self, resolvingAgainstBaseURL: false),
+              let queryItems = components.queryItems else {
+            return nil
+        }
+        return Dictionary(uniqueKeysWithValues: queryItems.map { ($0.name, $0.value ?? "") })
+    }
+}
